test/java/org/opendatakit/briefcase/reused/http/CommonsHttpTest.java
/*
 * Copyright (C) 2018 Nafundi
 *
 * Licensed under the Apache License, Version 2.0 (the "License"); you may not
 * use this file except in compliance with the License. You may obtain a copy of
 * the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
 * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 * License for the specific language governing permissions and limitations under
 * the License.
 */

package org.opendatakit.briefcase.reused.http;

import static org.hamcrest.Matchers.containsString;
import static org.junit.Assert.assertThat;

import org.junit.Test;

public class CommonsHttpTest {
  @Test
  public void can_execute_a_query() {
    Http http = new CommonsHttp();
    String output = http.execute(Query.normal("https://github.com/opendatakit/briefcase"));
    assertThat(output, containsString("ODK Briefcase is a Java application for fetching and pushing forms and their contents"));
  }
}